Rename `OrderInfo` to `index` and update import paths
