+++ react/OrderPlaced.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types'
 import { compose, graphql } from 'react-apollo'
 import { branch, renderComponent } from 'recompose'
 import { Loader } from 'vtex.order-details'
-import Header from './components/Header/Header'
-import OrderInfo from './components/OrderInfo/OrderInfo'
+import Header from './components/Header'
+import OrderInfo from './components/OrderInfo'
 import getOrderGroup from './graphql/getOrderGroup.graphql'
 import withoutSSR from './withoutSSR'
